Ensure we also set the QuicConnectionId when the id has a length of 0.
Motivation:

We did not correctly handle the case when the remote peer used a connection id with length of 0.

Modifications:

- Correctly handle the case when an id is of length 0
- Modify unit tests to catch the problem

Result:

Correct handling of zero length connection ids
