+++ codec-native-quic/src/test/java/io/netty/incubator/codec/quic/QuicChannelValidationHandler.java
@@ -18,8 +18,13 @@ package io.netty.incubator.codec.quic;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 
-class QuicChannelValidationHandler extends ChannelInboundHandlerAdapter {
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+class QuicChannelValidationHandler extends ChannelInboundHandlerAdapter {
+    private volatile boolean wasActive;
+
+    private volatile QuicConnectionAddress localAddress;
+    private volatile QuicConnectionAddress remoteAddress;
     private volatile Throwable cause;
 
     @Override
@@ -27,10 +32,30 @@ class QuicChannelValidationHandler extends ChannelInboundHandlerAdapter {
         this.cause = cause;
     }
 
+    @Override
+    public void channelActive(ChannelHandlerContext ctx) {
+        wasActive = true;
+        localAddress = (QuicConnectionAddress) ctx.channel().localAddress();
+        remoteAddress = (QuicConnectionAddress) ctx.channel().remoteAddress();
+        ctx.fireChannelActive();
+    }
+
+    QuicConnectionAddress localAddress() {
+        return localAddress;
+    }
+
+    QuicConnectionAddress remoteAddress() {
+        return remoteAddress;
+    }
+
     void assertState() throws Throwable {
         if (cause != null) {
             throw cause;
         }
+        if (wasActive) {
+            // Validate that the addresses could be retrieved
+            assertNotNull(localAddress);
+            assertNotNull(remoteAddress);
+        }
     }
-
 }
