+++ codec-native-quic/src/test/java/io/netty/incubator/codec/quic/QuicStreamFrameTest.java
@@ -83,6 +83,7 @@ public class QuicStreamFrameTest extends AbstractQuicTest {
 
         @Override
         public void channelActive(ChannelHandlerContext ctx) {
+            super.channelActive(ctx);
             QuicChannel channel = (QuicChannel) ctx.channel();
             channel.createStream(type, new ChannelInboundHandlerAdapter() {
                 @Override
